Add basic Routing
- Move content from App.tsx to ClassOverview.tsx
- Add Routing to App
